Use a single kernel and initramfs.cpio.gz. Signed-off-by: Trammell hudson <hudson@trmm.net>
Remove the `boot`/`full` kernel distinction; only build the
full kernel and launch it directly from the aspeed startup code.

This speeds up the boot process and reduces the amount of flash used.
The kernel and initramfs are also compressed with gz instead of xz,
since the ARM cpus are 10x faster at decompressing gz.